Fix flake8 invocation in .travis.yml
flake8 was invoked with '*.py tests' which was only checking
setup.py and Python files in tests/, and not checking any of
the nested_dict modules.

Use tox.ini to specify the files to be excluded, which allows
flake8 to include all other Python files found.

Fix whitespace related errors.

Ignore code E265 due to several `#print..` lines.

Also simplify the invocation of the sphinx docs command in .travis.yml